requirements.txt
Django>=4.2,<5.0
pyserial
websocket-client